refactor: forum service
