+++ __test__/response/forum_topic.json
@@ -1,0 +1,14 @@
+{
+  "category": "Genshin",
+  "cid": "QmQoKy9svWsjyKrpwTSEADc3FxamfFBDuYcXcG91D9jcZ6",
+  "createTime": "2023-07-13T07:30:02.987262099Z",
+  "creator": "User404830553902647325714607466232238944473254011640",
+  "id": "f7043b0f-bd58-4d3e-bc22-97c4a5559ce6",
+  "images": [
+    "QmXPwxdwh8RLikvkTDpCBKvJzCvo6CbsgugFx6Qu5JUnnk",
+    "Qma3U8HM5xUoKrhXveRr4KJFwZVJswiEntATLGnHCkMZWn"
+  ],
+  "tags": ["win", "lose"],
+  "title": "Genshin Impact, run!",
+  "updateTime": "2023-07-13T07:30:02.987262099Z"
+}
