+++ babel.config.js
@@ -9,6 +9,7 @@ module.exports = function (api) {
           root: ["."],
           alias: {
             "@src": "./src",
+            "@root": ".",
           },
         },
       ],
