+++ src/utils/numericCarry.ts
@@ -4,5 +4,5 @@ export function numericCarry(num: number): string {
   if (tmp > 99) return "99k+";
   const str = String(tmp * 10);
   const match = str.match(/^\d+$|^\d+\./) as RegExpExecArray;
-  return parseFloat(match[0]) / 10.0 + "k";
+  return `${parseFloat(match[0]) / 10.0}k`;
 }
